backend/requirements.txt
logger==1.4
pydantic==2.10.1
pydantic_core==2.27.1
pydantic-settings==2.6.1
python-doctr==0.10.0
ultralytics==8.3.36
python-multipart==0.0.17
uvicorn==0.32.1
fastapi==0.115.5
opencv-python-headless